requirements.txt
python-barcode>=0.16.1
Pillow>=12.2.0